src/printforge/multi_view.py
"""Multi-View Enhancement: Single image → multiple views for better 3D reconstruction."""

import logging
from dataclasses import dataclass
from typing import Optional

from PIL import Image

logger = logging.getLogger(__name__)


@dataclass
class MultiViewConfig:
    """Configuration for multi-view generation."""
    num_views: int = 6
    # Azimuth angles in degrees for each view
    azimuths: tuple = (0, 60, 120, 180, 240, 300)
    elevation: float = 20.0
    image_size: int = 512
    # Model backend: "placeholder", "zero123pp", "sv3d"
    backend: str = "placeholder"


class MultiViewEnhancer:
    """Generate multiple views from a single image.

    Currently implements a placeholder that returns copies of the original image.
    Future implementations will use Zero123++ or SV3D for novel view synthesis.
    """

    def __init__(self, config: Optional[MultiViewConfig] = None):
        self.config = config or MultiViewConfig()

    def generate_views(self, image_path: str) -> list[Image.Image]:
        """Generate multiple views from a single input image.

        Args:
            image_path: Path to the input image.

        Returns:
            List of PIL Images representing different views.
        """
        if self.config.backend == "placeholder":
            return self._placeholder_views(image_path)
        elif self.config.backend == "zero123pp":
            return self._zero123pp_views(image_path)
        elif self.config.backend == "sv3d":
            return self._sv3d_views(image_path)
        else:
            raise ValueError(f"Unknown backend: {self.config.backend}")

    def save_views(self, views: list[Image.Image], output_dir: str) -> list[str]:
        """Save generated views to disk.

        Returns list of saved file paths.
        """
        from pathlib import Path

        out = Path(output_dir)
        out.mkdir(parents=True, exist_ok=True)

        paths = []
        for i, view in enumerate(views):
            azimuth = self.config.azimuths[i] if i < len(self.config.azimuths) else i * 60
            path = out / f"view_{azimuth:03d}.png"
            view.save(str(path))
            paths.append(str(path))

        logger.info(f"Saved {len(paths)} views to {output_dir}")
        return paths

    def _placeholder_views(self, image_path: str) -> list[Image.Image]:
        """Placeholder: return copies of the original image for each view angle."""
        img = Image.open(image_path).convert("RGB")
        img = img.resize((self.config.image_size, self.config.image_size), Image.LANCZOS)

        views = []
        for i in range(self.config.num_views):
            views.append(img.copy())

        logger.info(f"Placeholder: generated {len(views)} view copies from {image_path}")
        return views

    def _zero123pp_views(self, image_path: str) -> list[Image.Image]:
        """Zero123++ novel view synthesis. Requires zero123plus package."""
        raise NotImplementedError(
            "Zero123++ backend not yet implemented. "
            "Install zero123plus and implement diffusion-based view synthesis."
        )

    def _sv3d_views(self, image_path: str) -> list[Image.Image]:
        """SV3D novel view synthesis. Requires sv3d package."""
        raise NotImplementedError(
            "SV3D backend not yet implemented. "
            "Install sv3d and implement video-diffusion-based view synthesis."
        )
